CamelCase
Alice wrote a sequence of words in CamelCase as a string of letters, s, having the following properties:
It is a concatenation of one or more words consisting of English letters.
All letters in the first word are lowercase.

For each of the subsequent words, the first letter is uppercase and rest of the letters are lowercase.
Given s, print the number of words in s on a new line.
For example, s=oneTwoThree. There are 3 words in the string.

Function Description
Complete the camelcase function in the editor below. It must return the integer number of words in the input string.
camelcase has the following parameter(s):
s: the string to analyze

Input Format
A single line containing string s.